Remove iconv from go-charset
According to the description in the go-charset's project page, iconv is an
extra package, which is assumed to be an optional package.  To remove
dependency from cgo, remove the iconv from go-charset.

The removal presumably will lessen the number of supported charset. But, let's
worry about it later :)

Now the whole package is testable using Go bundled with Google App Engine.

Closes #14